+++ CHANGES.md
@@ -1,7 +1,9 @@
 # Changes
 
-## v0.5.1
+## v0.6.0
 
+- Drafts: New files have timetamps set to 'DRAFT' initially. Once blogdown gets
+  invoked with `--publish` the timestamps are set to that moment.
 - JSON: Support newlines and indentation in strings
 - JSON: Resolve multiple placeholders in single property
 - Fix: Remove dots from file names if they are generated from another property